// job-tracker-ui/src/types/auth.ts
export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  email: string;
  password: string;
  displayName: string;
}

export interface AuthResponse {
  token: string;
  email: string;
  displayName: string;
}

export interface AuthUser {
  email: string;
  displayName: string;
}